+++ packages/toolkit/utils/src/wait.ts
@@ -1,6 +1,0 @@
-const wait = (time = 0) =>
-  new Promise(resolve => {
-    setTimeout(resolve, time);
-  });
-
-export { wait };