+++ packages/toolkit/utils/tests/logger.test.ts
@@ -1,4 +1,4 @@
-import { logger } from '../src/logger';
+import { logger } from '../src';
 
 describe('logger', () => {
   test('should log', () => {