+++ tests/utils/setup.js
@@ -7,11 +7,13 @@ const puppeteer = require('puppeteer');
 const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
 
 module.exports = async function () {
-  const browser = await puppeteer.launch({ args: ['--no-sandbox'] });
+  const browser = await puppeteer.launch({
+    args: ['--no-sandbox'],
+    // headless: false,
+  });
   // store the browser instance so we can teardown it later
   // this global is only available in the teardown but not in TestEnvironments
   global.__BROWSER_GLOBAL__ = browser;
-  global.page = await browser.newPage();
 
   // use the file system to expose the wsEndpoint for TestEnvironments
   mkdirp.sync(DIR);