+++ packages/builder/builder-shared/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/builder-shared
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/server@2.18.0
+- @modern-js/types@2.18.0
+- @modern-js/utils@2.18.0
+
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/server@2.17.1
+- @modern-js/types@2.17.1
+- @modern-js/utils@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes