+++ packages/cli/plugin-lint/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/plugin-lint
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.18.0
+- @modern-js/utils@2.18.0
+
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.17.1
+- @modern-js/utils@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes