+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/generator-utils
 
+## 3.1.13
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.18.0
+- @modern-js/utils@2.18.0
+- @modern-js/generator-common@3.1.13
+
+## 3.1.12
+
+### Patch Changes
+
+- @modern-js/generator-common@3.1.12
+- @modern-js/plugin-i18n@2.17.1
+- @modern-js/utils@2.17.1
+
 ## 3.1.11
 
 ### Patch Changes