+++ packages/generator/generators/upgrade-generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/upgrade-generator
 
+## 3.1.13
+
+## 3.1.12
+
 ## 3.1.11
 
 ## 3.1.10