+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/server-utils
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/babel-preset-lib@2.18.0
+- @modern-js/babel-compiler@2.18.0
+- @modern-js/utils@2.18.0
+
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/babel-preset-lib@2.17.1
+- @modern-js/babel-compiler@2.17.1
+- @modern-js/utils@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes