+++ scripts/prebundle/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scripts/prebundle
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.18.0
+- @scripts/build@2.18.0
+
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.17.1
+- @scripts/build@2.17.1
+
 ## 2.17.0
 
 ### Patch Changes