+++ scripts/update-codesmith/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scripts/update-codesmith
 
+## 2.18.0
+
+### Patch Changes
+
+- @modern-js/generator-utils@3.1.13
+
+## 2.17.1
+
+### Patch Changes
+
+- @modern-js/generator-utils@3.1.12
+
 ## 2.17.0
 
 ### Patch Changes